+++ packages/block-editor/src/utils/test/get-font-styles-and-weights.js
@@ -34,10 +34,6 @@ describe( 'getFontStylesAndWeights', () => {
 					value: '400',
 				},
 				{
-					name: 'Medium',
-					value: '500',
-				},
-				{
 					name: 'Semi Bold',
 					value: '600',
 				},
@@ -89,14 +85,6 @@ describe( 'getFontStylesAndWeights', () => {
 					style: {
 						fontStyle: 'normal',
 						fontWeight: '400',
-					},
-				},
-				{
-					key: 'normal-500',
-					name: 'Medium',
-					style: {
-						fontStyle: 'normal',
-						fontWeight: '500',
 					},
 				},
 				{
@@ -172,14 +160,6 @@ describe( 'getFontStylesAndWeights', () => {
 					},
 				},
 				{
-					key: 'italic-500',
-					name: 'Medium Italic',
-					style: {
-						fontStyle: 'italic',
-						fontWeight: '500',
-					},
-				},
-				{
 					key: 'italic-600',
 					name: 'Semi Bold Italic',
 					style: {
@@ -252,10 +232,6 @@ describe( 'getFontStylesAndWeights', () => {
 					value: '400',
 				},
 				{
-					name: 'Medium',
-					value: '500',
-				},
-				{
 					name: 'Semi Bold',
 					value: '600',
 				},
@@ -271,14 +247,6 @@ describe( 'getFontStylesAndWeights', () => {
 					style: {
 						fontStyle: 'normal',
 						fontWeight: '400',
-					},
-				},
-				{
-					key: 'normal-500',
-					name: 'Medium',
-					style: {
-						fontStyle: 'normal',
-						fontWeight: '500',
 					},
 				},
 				{
@@ -303,14 +271,6 @@ describe( 'getFontStylesAndWeights', () => {
 					style: {
 						fontStyle: 'italic',
 						fontWeight: '400',
-					},
-				},
-				{
-					key: 'italic-500',
-					name: 'Medium Italic',
-					style: {
-						fontStyle: 'italic',
-						fontWeight: '500',
 					},
 				},
 				{